ggml: use int64_t for k in dequantize_row_q4_hqq signature
Update the `dequantize_row_q4_hqq()` function signature to use `int64_t`
for the `k` parameter instead of `int`.

This aligns the function with the conventions used by other ggml
quantization/dequantization routines (e.g., `dequantize_row_q4_0`,
`dequantize_row_q4_1`, etc.), which consistently use `int64_t` for
element counts.

Using `int64_t` ensures type consistency across the ggml quantization API
and avoids potential issues when working with large tensors where the
element count may exceed the range of a 32-bit integer.

No functional changes to the dequantization logic.

+++ ggml/src/ggml-quants.c
@@ -388,7 +388,7 @@ void dequantize_row_q4_1(const block_q4_1 * GGML_RESTRICT x, float * GGML_RESTRI
     }
 }
 
-void dequantize_row_q4_hqq(const block_q4_hqq * x, float * y, int k) {
+void dequantize_row_q4_hqq(const block_q4_hqq * x, float * y, int64_t k) {
 
     const int nb = k / QK4_HQQ;
